# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "這是一個輕量級的微服務範本，適合快速啟動專案。它支援 In-Memory 和 MySQL 兩種模式。",
        "title": "Microservice MVP API (範本)",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/login": {
            "post": {
                "description": "驗證玩家憑證並返回認證 Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "玩家登入",
                "parameters": [
                    {
                        "description": "登入請求參數",
                        "name": "loginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/microservice-mvp_internal_model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登入成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/microservice-mvp_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/microservice-mvp_internal_model.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "請求參數錯誤",
                        "schema": {
                            "$ref": "#/definitions/microservice-mvp_pkg_response.HTTPError400"
                        }
                    },
                    "401": {
                        "description": "認證失敗",
                        "schema": {
                            "$ref": "#/definitions/microservice-mvp_pkg_response.HTTPError401"
                        }
                    }
                }
            }
        },
        "/api/v1/players/{id}": {
            "get": {
                "description": "根據玩家 ID 取得玩家的詳細資料，包含餘額",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Player"
                ],
                "summary": "取得玩家資料",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "玩家 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功取得玩家資料",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/microservice-mvp_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/microservice-mvp_internal_model.PlayerInfoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "請求參數錯誤",
                        "schema": {
                            "$ref": "#/definitions/microservice-mvp_pkg_response.HTTPError400"
                        }
                    },
                    "404": {
                        "description": "玩家不存在",
                        "schema": {
                            "$ref": "#/definitions/microservice-mvp_pkg_response.HTTPError404"
                        }
                    },
                    "500": {
                        "description": "內部伺服器錯誤",
                        "schema": {
                            "$ref": "#/definitions/microservice-mvp_pkg_response.HTTPError500"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "檢查服務運行狀態、系統指標及依賴組件健康狀況",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "健康檢查",
                "responses": {
                    "200": {
                        "description": "服務正常",
                        "schema": {
                            "$ref": "#/definitions/internal_controller.HealthCheckResponse"
                        }
                    },
                    "503": {
                        "description": "服務降級或異常",
                        "schema": {
                            "$ref": "#/definitions/internal_controller.HealthCheckResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_controller.ComponentStatus": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "額外資訊 (例如: 項目數量, DB 統計)",
                    "type": "string"
                },
                "latency": {
                    "type": "string",
                    "example": "5ms"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "UP"
                }
            }
        },
        "internal_controller.HealthCheckResponse": {
            "type": "object",
            "properties": {
                "components": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/internal_controller.ComponentStatus"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "UP"
                },
                "system": {
                    "$ref": "#/definitions/internal_controller.SystemMetrics"
                },
                "uptime": {
                    "type": "string",
                    "example": "1h2m3s"
                }
            }
        },
        "internal_controller.SystemMetrics": {
            "type": "object",
            "properties": {
                "go_version": {
                    "type": "string",
                    "example": "go1.23.0"
                },
                "goroutines": {
                    "type": "integer",
                    "example": 12
                },
                "memory_usage": {
                    "type": "string",
                    "example": "5 MB"
                }
            }
        },
        "microservice-mvp_internal_model.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "microservice-mvp_internal_model.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "JWT Token 的佔位符",
                    "type": "string"
                }
            }
        },
        "microservice-mvp_internal_model.PlayerInfoResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "microservice-mvp_pkg_response.HTTPError400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "請求參數錯誤"
                }
            }
        },
        "microservice-mvp_pkg_response.HTTPError401": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "message": {
                    "type": "string",
                    "example": "未經授權"
                }
            }
        },
        "microservice-mvp_pkg_response.HTTPError404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "message": {
                    "type": "string",
                    "example": "找不到資源"
                }
            }
        },
        "microservice-mvp_pkg_response.HTTPError500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "type": "string",
                    "example": "內部伺服器錯誤"
                }
            }
        },
        "microservice-mvp_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    }
}